// packages/mermaid/src/config.type.d.ts
export interface MermaidConfig {
  /**
   * Theme, the CSS style sheet.
   * You may also use `themeCSS` to override this value.
   *
   */
  theme?:
    | 'default'
    | 'base'
    | 'dark'
    | 'forest'
    | 'neutral'
    | 'neo'
    | 'neo-dark'
    | 'redux'
    | 'redux-dark'
    | 'redux-color'
    | 'redux-dark-color'
    | 'null';
  themeVariables?: any;
  themeCSS?: string;
  /**
   * Defines which main look to use for the diagram.
   *
   */
  look?: 'classic' | 'handDrawn' | 'neo';
// ...
}